+++ zucchini-ui-backend/src/main/java/io/zucchiniui/backend/scenario/domain/ScenarioReviewedStateChange.java
@@ -1,0 +1,17 @@
+package io.zucchiniui.backend.scenario.domain;
+
+import java.time.ZonedDateTime;
+
+public class ScenarioReviewedStateChange extends ScenarioChange<Boolean> {
+
+    /**
+     * Private constructor for Morphia.
+     */
+    private ScenarioReviewedStateChange() {
+    }
+
+    public ScenarioReviewedStateChange(ZonedDateTime date, boolean oldState, boolean newState) {
+        super(ChangeType.REVIEWED_STATE, date, oldState, newState);
+    }
+
+}