+++ zucchini-ui-frontend/app/scripts/tc-scenario-change.js
@@ -1,0 +1,13 @@
+'use strict';
+
+var zucchiniModule = require('./module');
+
+
+zucchiniModule
+  .component('tcScenarioChange', {
+    template: require('../views/tc-scenario-change.html'),
+    bindings: {
+      type: '<',
+      value: '<',
+    },
+  });